style: satisfy cyber session blocking lint checks

+++ internal/server/biz/system.go
@@ -31,7 +31,7 @@ import (
 
 const (
 	maxRetryResponseTimeoutSeconds = 600
-	maxCyberSessionBlockTTLSeconds = int64((1<<63 - 1) / int64(time.Second))
+	maxCyberSessionBlockTTLSeconds = (1<<63 - 1) / int64(time.Second)
 )
 
 const (
